:recycle: Extract permissions checker and improve nested entity permission checks on querying
